refactor: change structure project and edit state
