+++ src/components/ActiveQuiz/Answer/AnswerDetails/AnswerDescription/AnswerDescription.css
@@ -1,23 +1,3 @@
-.answer-details {
-    max-width: 100%;
-}
-
-.answer-details-card {
-    display: flex;
-    padding: 1rem;
-    border: 1px solid #444;
-    max-width: 100%;
-    min-height: 315px;
-}
-
-.answer-details .list-group-item {
-    padding: .25rem;
-}
-
-.answer-details .list-group-item .term {
-    margin-right: .5rem;
-}
-
 .card-body {
     display: flex;
     padding: 0;
@@ -27,12 +7,6 @@
 
 .card-body .list-group {
     width: calc(100% - 215px);
-}
-
-@media (max-width: 1200px) {
-    .answer-details-card {
-        min-height: 333px;
-    }
 }
 
 @media (max-width: 992px) {
