+++ src/components/ActiveQuiz/Answer/AnswerDetails/AnswerDetails.js
@@ -1,7 +1,6 @@
-import React, { Fragment } from 'react';
+import React from 'react';
 import './AnswerDetails.css';
-import AudioPlayer from 'react-h5-audio-player';
-import './../Player/Player.scss';
+import { AnswerDescription } from './AnswerDescription/AnswerDescription';
 
 export const AnswerDetails = (props) => {
   return (
@@ -10,27 +9,13 @@ export const AnswerDetails = (props) => {
         { props.answerId
           ? props.info.map(key => {
             if(key.id === props.answerId) {
-              return <Fragment key={ key.id }>
-                <div className='card-body'>
-                  <img className='bird-image' src={ key.image } alt="Bird" />
-                  <ul className="list-group list-group-flush">
-                    <li className="list-group-item">
-                      <h4>{ key.name }</h4>
-                    </li>
-                    <li className="list-group-item">{ key.song }</li>
-                    <li className="list-group-item">
-                      <AudioPlayer
-                        autoPlayAfterSrcChange={ false }
-                        volume='0.5'
-                        layout='stacked'
-                        showJumpControls={ true }
-                        src={ key.audio }
-                      />
-                    </li>
-                  </ul>
-                </div>
-                <span className='card-text'>{ key.description }</span>
-              </Fragment>;
+              return <AnswerDescription
+                key={ key.id }
+                srcImage={ key.image }
+                name={ key.name }
+                song={ key.song }
+                srcAudio={ key.audio }
+                description={key.description} />
             }
           })
           : <div><h3>Правила игры</h3>
